feat: Allow the user to select a program
Added a program list the user can search through,
and added a method to append/remove elements to the list.
For this, list_elements was modified to add a search bar.
We also created a program class which can obtain their information.
It can be a bit slow though, so we still need to optimize adding
thousands of items to the list.