feat: Create a thread-safe cache class which will be used in the near future.
This class is thread-safe and can be used in a multi-threaded environment.
It also ensure that the cache cannot get corrupted during a write operation.
It will be used to store the results of different operations, such as obtaining the program history.
This will allow to avoid computing this expensive operation multiple times.