fix: Several bugfixes
- Bugfix: We couldn't find the different app modes in certain cases

If you have multiple monitors, or too slow of a CPU, the app modes would not get detected, rendering many features unusable.

- Bugfix: We couldn't find the channel list in certain cases

See above

- Bugfix: Thread safety issue with list_elements

Threads appending elements to the list would not stop if the list was closed, crashing NVDA

- Bugfix: Empty lists with list_elements:

If a list of element empty, the callback would still be called with the empty list placeholder, causing unexpected behavior.

- Bugfix: Sometimes scrolling down to find a program would stop halfway through if your screen was too small.

Added a hovering offset for this.

- Bugfix: Certain elements don't match their names in the element list.

This was due to a bug in the way we were getting the element names in the appendElements function. This has been fixed.